refactor Card & Position components to follow DRY
