+++ src/pages/SquadBuilder/Position/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Position'
